final_mission_code/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain 생태계
langchain==0.2.16
langgraph==0.2.28
langchain-core==0.2.40
langchain-community==0.2.16

# AI 모델 통합
langchain-anthropic==0.1.23
langchain-aws==0.1.17
langchain-upstage==0.1.8
langchain-huggingface==0.0.3
langchain-tavily==0.1.1

# 벡터 데이터베이스
langchain-postgres==0.0.9
psycopg2-binary==2.9.9
pgvector==0.2.5

# 임베딩 및 ML
sentence-transformers==2.7.0
torch==2.1.0
transformers==4.36.0

# 유틸리티
python-dotenv==1.0.0
pydantic==2.5.0

# 모니터링 및 추적
langfuse==2.36.0

# 개발 도구
pytest==7.4.3
black==23.11.0